requirements.txt
# Dev environment
pip
autopep8

# App
Pillow # PIL image-loading library
python-fasthtml